automation_client/requirements.txt
requests>=2.28.0
pystray>=0.19.0
Pillow>=10.0.0
pyinstaller>=6.0.0
